Re-write the edit section.
See:
https://github.com/ppy/osu/pull/28806/files
